requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiofiles==23.2.1
python-multipart==0.0.9
